+++ src/main/java/com/emergency/challenge/TalkApplication.java
@@ -1,0 +1,13 @@
+package com.emergency.challenge;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TalkApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(TalkApplication.class, args);
+    }
+
+}
