+++ src/main/java/com/emergency/challenge/config/JwtSecurityConfiguration.java
@@ -1,0 +1,25 @@
+package com.emergency.challenge.config;
+
+import com.emergency.challenge.jwt.JwtFilter;
+import com.emergency.challenge.jwt.TokenProvider;
+import com.emergency.challenge.service.UserDetailsServiceImpl;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.config.annotation.SecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.web.DefaultSecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+@RequiredArgsConstructor
+public class JwtSecurityConfiguration
+        extends SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity> {
+
+    private final String SECRET_KEY;
+    private final TokenProvider tokenProvider;
+    private final UserDetailsServiceImpl userDetailsService;
+
+    @Override
+    public void configure(HttpSecurity httpSecurity) {
+        JwtFilter customJwtFilter = new JwtFilter(SECRET_KEY, tokenProvider, userDetailsService);
+        httpSecurity.addFilterBefore(customJwtFilter, UsernamePasswordAuthenticationFilter.class);
+    }
+}
