+++ src/main/java/com/emergency/challenge/controller/CustomBaseControllerAnnotation.java
@@ -1,0 +1,15 @@
+package com.emergency.challenge.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@RestController
+@RequestMapping("/api/*")
+public @interface CustomBaseControllerAnnotation {}
