+++ src/main/java/com/emergency/challenge/controller/response/ResponseDto.java
@@ -1,0 +1,30 @@
+package com.emergency.challenge.controller.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@AllArgsConstructor
+public class ResponseDto<T> {
+    private boolean success;
+    private T data;
+    private Error error;
+
+    public static <T> ResponseDto<T> success(T data) {
+
+        return new ResponseDto<>(true, data, null);
+    }
+
+    public static <T> ResponseDto<T> fail(String code, String message) {
+        return new ResponseDto<>(false, null, new Error(code, message));
+    }
+
+    @Getter
+    @Setter
+    @AllArgsConstructor
+    static class Error {
+        private String code;
+        private String message;
+    }
+}
