+++ src/main/java/com/emergency/challenge/repository/MemberRepository.java
@@ -1,0 +1,11 @@
+package com.emergency.challenge.repository;
+
+import com.emergency.challenge.domain.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    Optional<Member> findById(Long id);
+    Optional<Member> findByNickname(String nickname);
+}
