+++ src/main/java/com/emergency/challenge/repository/RefreshTokenRepository.java
@@ -1,0 +1,11 @@
+package com.emergency.challenge.repository;
+
+import com.emergency.challenge.domain.Member;
+import com.emergency.challenge.domain.RefreshToken;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+    Optional<RefreshToken> findByMember(Member member);
+}
