+++ src/main/java/com/emergency/challenge/shared/Authority.java
@@ -1,0 +1,7 @@
+package com.emergency.challenge.shared;
+
+public enum Authority {
+    ROLE_ADMIN,
+    ROLE_MEMBER,
+    ROLE_GUEST
+}
