--- src/main/java/com/emergency/talk/TalkApplication.java
@@ -1,13 +1,0 @@
-package com.emergency.talk;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TalkApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(TalkApplication.class, args);
-    }
-
-}
