--- src/main/java/com/emergency/talk/domain/RefreshToken.java
@@ -1,31 +1,0 @@
-package com.emergency.talk.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Entity
-public class RefreshToken extends Timestamped {
-
-    @Id
-    @Column(nullable = false)
-    private Long id;
-
-    @JoinColumn(name = "member_id", nullable = false)
-    @OneToOne(fetch = FetchType.LAZY)
-    private Member member;
-
-    @Column(nullable = false)
-    private String value;
-
-    public void updateValue(String token) {
-        this.value = token;
-    }
-}
