--- src/main/java/com/emergency/talk/error/RestApiException.java
@@ -1,14 +1,0 @@
-package com.emergency.talk.error;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class RestApiException {
-    private String field;
-    private String errorMessage;
-
-}
