--- src/main/java/com/emergency/talk/jwt/AccessDeniedHandlerException.java
@@ -1,29 +1,0 @@
-package com.emergency.talk.jwt;
-
-import com.emergency.talk.controller.response.ResponseDto;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@Component
-public class AccessDeniedHandlerException implements AccessDeniedHandler {
-
-    @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response,
-                       AccessDeniedException accessDeniedException) throws IOException {
-
-        response.setContentType("application/json;charset=UTF-8");
-        response.getWriter().println(
-                new ObjectMapper().writeValueAsString(
-                        ResponseDto.fail("BAD_REQUEST", "로그인 해주시길 바래요.")
-                )
-        );
-
-        response.setStatus(HttpServletResponse.SC_FORBIDDEN);
-    }
-}
