--- src/main/java/com/emergency/talk/repository/MemberRepository.java
@@ -1,11 +1,0 @@
-package com.emergency.talk.repository;
-
-import com.emergency.talk.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findById(Long id);
-    Optional<Member> findByNickname(String nickname);
-}
