--- src/main/java/com/emergency/talk/repository/RefreshTokenRepository.java
@@ -1,11 +1,0 @@
-package com.emergency.talk.repository;
-
-import com.emergency.talk.domain.Member;
-import com.emergency.talk.domain.RefreshToken;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
-    Optional<RefreshToken> findByMember(Member member);
-}
