--- src/main/java/com/emergency/talk/shared/Authority.java
@@ -1,7 +1,0 @@
-package com.emergency.talk.shared;
-
-public enum Authority {
-    ROLE_ADMIN,
-    ROLE_MEMBER,
-    ROLE_GUEST
-}
