+++ src/main/resources/message.properties
@@ -1,4 +1,4 @@
 nickname.notblank=Id는 공백일 수 없습니다.!
-nickname.option=id형식 이상, 4 ~ 12 로 하셈.!
+nickname.option=id형식 이상, 4 ~ 12 로 하세요.!
 password.option=패스워드는 무조건 영문, 숫자, 특수문자를 각각 1글자 이상 포함해야 합니다.!
 password.notblank=패스워드는 공백일 수 없습니다.!
