+++ src/test/java/com/emergency/challenge/TalkApplicationTests.java
@@ -1,0 +1,13 @@
+package com.emergency.challenge;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TalkApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
