--- src/test/java/com/emergency/talk/TalkApplicationTests.java
@@ -1,13 +1,0 @@
-package com.emergency.talk;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TalkApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
